oracle proof producer instead of submitter
