+++ prover/proof_submitter/util_test.go
@@ -2,17 +2,11 @@ package submitter
 
 import (
 	"context"
-	"crypto/ecdsa"
-	"crypto/elliptic"
 	"errors"
 	"math/big"
-	"math/rand"
-	"os"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/taikoxyz/taiko-client/bindings/encoding"
 )
 
 func (s *ProofSubmitterTestSuite) TestIsSubmitProofTxErrorRetryable() {
@@ -57,51 +51,4 @@ func (s *ProofSubmitterTestSuite) TestSendTxWithBackoff() {
 	})
 
 	s.Nil(err)
-}
-
-// randomHash generates a random blob of data and returns it as a hash.
-func randomHash() common.Hash {
-	var hash common.Hash
-	if n, err := rand.Read(hash[:]); n != common.HashLength || err != nil {
-		panic(err)
-	}
-	return hash
-}
-
-func (s *ProofSubmitterTestSuite) TestHashAndSignEvidenceForOracleProof() {
-	evidence := &encoding.TaikoL1Evidence{
-		MetaHash:      randomHash(),
-		BlockHash:     randomHash(),
-		ParentHash:    randomHash(),
-		SignalRoot:    randomHash(),
-		Graffiti:      randomHash(),
-		Prover:        common.BigToAddress(new(big.Int).SetUint64(rand.Uint64())),
-		ParentGasUsed: 1024,
-		GasUsed:       1024,
-		VerifierId:    0,
-		Proof:         nil,
-	}
-
-	privateKey, err := crypto.HexToECDSA(os.Getenv("L1_PROVER_PRIVATE_KEY"))
-	s.Nil(err)
-
-	publicKey := privateKey.Public()
-
-	publicKeyECDSA, ok := publicKey.(*ecdsa.PublicKey)
-	s.True(ok)
-
-	input, err := encoding.EncodeProveBlockInput(evidence)
-	s.Nil(err)
-
-	hash := crypto.Keccak256Hash(input)
-
-	sig, _, err := hashSignAndSetEvidenceForOracleProof(evidence, privateKey)
-	s.Nil(err)
-
-	pubKey, err := crypto.Ecrecover(hash.Bytes(), sig)
-	s.Nil(err)
-	isValid := crypto.VerifySignature(pubKey, hash.Bytes(), sig[:64])
-	s.True(isValid)
-
-	s.Equal(elliptic.Marshal(publicKeyECDSA, publicKeyECDSA.X, publicKeyECDSA.Y), pubKey)
 }
