+++ prover/prover.go
@@ -123,47 +123,44 @@ func InitFromConfig(ctx context.Context, p *Prover, cfg *Config) (err error) {
 	p.proposeConcurrencyGuard = make(chan struct{}, cfg.MaxConcurrentProvingJobs)
 	p.submitProofConcurrencyGuard = make(chan struct{}, cfg.MaxConcurrentProvingJobs)
 
+	var producer proofProducer.ProofProducer
+
 	if cfg.OracleProver {
-		// Proof submitter
-		if p.validProofSubmitter, err = proofSubmitter.NewOracleProofSubmitter(
+		if producer, err = proofProducer.NewOracleProducer(
 			p.rpc,
-			p.proveValidProofCh,
-			p.cfg.TaikoL2Address,
 			p.cfg.L1ProverPrivKey,
-			p.submitProofTxMutex,
+			p.cfg.TaikoL2Address,
 		); err != nil {
 			return err
+		}
+	} else if cfg.Dummy {
+		producer = &proofProducer.DummyProofProducer{
+			RandomDummyProofDelayLowerBound: p.cfg.RandomDummyProofDelayLowerBound,
+			RandomDummyProofDelayUpperBound: p.cfg.RandomDummyProofDelayUpperBound,
 		}
 	} else {
-		var producer proofProducer.ProofProducer
-
-		if cfg.Dummy {
-			producer = &proofProducer.DummyProofProducer{
-				RandomDummyProofDelayLowerBound: p.cfg.RandomDummyProofDelayLowerBound,
-				RandomDummyProofDelayUpperBound: p.cfg.RandomDummyProofDelayUpperBound,
-			}
-		} else {
-			if producer, err = proofProducer.NewZkevmRpcdProducer(
-				cfg.ZKEvmRpcdEndpoint,
-				cfg.ZkEvmRpcdParamsPath,
-				cfg.L1HttpEndpoint,
-				cfg.L2HttpEndpoint,
-				true,
-			); err != nil {
-				return err
-			}
-		}
-		// Proof submitter
-		if p.validProofSubmitter, err = proofSubmitter.NewValidProofSubmitter(
-			p.rpc,
-			producer,
-			p.proveValidProofCh,
-			p.cfg.TaikoL2Address,
-			p.cfg.L1ProverPrivKey,
-			p.submitProofTxMutex,
+		if producer, err = proofProducer.NewZkevmRpcdProducer(
+			cfg.ZKEvmRpcdEndpoint,
+			cfg.ZkEvmRpcdParamsPath,
+			cfg.L1HttpEndpoint,
+			cfg.L2HttpEndpoint,
+			true,
 		); err != nil {
 			return err
 		}
+	}
+
+	// Proof submitter
+	if p.validProofSubmitter, err = proofSubmitter.NewValidProofSubmitter(
+		p.rpc,
+		producer,
+		p.proveValidProofCh,
+		p.cfg.TaikoL2Address,
+		p.cfg.L1ProverPrivKey,
+		p.submitProofTxMutex,
+		p.cfg.OracleProver,
+	); err != nil {
+		return err
 	}
 
 	return nil
